http2: fix unneeded cast found by clippy
error: casting to the same type is unnecessary (`u16` -> `u16`)
    --> src/http2/http2.rs:1286:47
     |
1286 |     if !orig_state.is_null() && orig_proto == ALPROTO_HTTP1 as u16 {
     |                                               ^^^^^^^^^^^^^^^^^^^^ help: try: `ALPROTO_HTTP1`
     |
     = help: for further information visit https://rust-lang.github.io/rust-clippy/rust-1.95.0/index.html#unnecessary_cast
note: the lint level is defined here
    --> src/lib.rs:23:38
     |
  23 | #![cfg_attr(feature = "strict", deny(warnings))]
     |                                      ^^^^^^^^
     = note: `#[deny(clippy::unnecessary_cast)]` implied by `#[deny(warnings)]`

+++ rust/src/http2/http2.rs
@@ -1283,7 +1283,7 @@ pub extern "C" fn rs_http2_state_new(
     let state = HTTP2State::new();
     let boxed = Box::new(state);
     let r = Box::into_raw(boxed) as *mut _;
-    if !orig_state.is_null() && orig_proto == ALPROTO_HTTP1 as u16 {
+    if !orig_state.is_null() && orig_proto == ALPROTO_HTTP1 {
         unsafe {
             HTTP2MimicHttp1Request(orig_state, r);
         }
